+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Rewards.hs
@@ -817,7 +817,7 @@ mkSnapShot ::
   SnapShot
 mkSnapShot activeStake delegs stakePools =
   SnapShot
-    { ssStake = activeStake
+    { ssActiveStake = activeStake
     , ssTotalActiveStake = totalActiveStake
     , ssDelegations = delegs
     , ssPoolParams = stakePools